# RFC: MapOf: custom comparators and better custom hashers