--- a/export_test.go
+++ b/export_test.go
@@ -70,11 +70,3 @@ func NextPowOf2(v uint32) uint32 {
 func MakeSeed() uint64 {
 	return makeSeed()
 }
-
-func HashString(s string, seed uint64) uint64 {
-	return hashString(s, seed)
-}
-
-func DefaultHasher[T comparable]() func(T, uint64) uint64 {
-	return defaultHasher[T]()
-}