This PR consists of two parts, one of which is probably less controversial and the other one is likely extremely controversial (hence the RFC mark).

I'm unfamiliar with Go and this ecosystem feels very alien to me, so please excuse me if any of this makes no sense.

1. Add support for custom comparators to `MapOf` in addition to custom hashers.
   As it happens, `xsync.MapOf` appears to be the best general-purpose hash map
   implementation that does not just defer to the built-in Go map. Thus, it has
   potential to support fully arbitrary (non-comparable) types, such as slices.
   Thus, make this potential a reality by adding support for custom comparators
   in addition to custom hashers.
    - **MapOf: allow a custom comparator in addition to a custom hasher**

2. Export some functions to simplify creating custom hashers.  
   Basically, this just re-exports xsync's wrappers over Go `runtime.*` hashers
   that are (ab)used internally. This is likely controversial, but I don't see
   a better idea. If/when Go finally decides to tighten the screws and strictly
   prohibit anyone from calling into `runtime.*`, both xsync's internal wrappers
   and exported wrappers can just be rewritten all at once, saving other projects
   the pain.
    - **util: define/use a `runtime_memhash64` similar to `runtime_typehash64`**
    - **util: export functions for creating custom hashers**