@puzpuzpuz 2025-03-31
v4 release will remove hasher API and migrate to the standard library `maphash.Comparable` function. This function is on-par with the private APIs (`runtime.typehash64`, `runtime.memhash64` and friends), so it makes little sense to keep using them. Considering this, I'm closing this PR. Feel free to reopen and/or comment if you have any v4 enhancement ideas.