@PapaCharlie 2025-03-31
So for the record, I played around with this idea because I was curious what it would look like. I relaxed the type bounds of `Map` and added a new way to create the map with custom hash and equality funcs. Changes can be found [here](https://github.com/PapaCharlie/xsync/compare/computev2.2...PapaCharlie:xsync:hasheq?expand=1). Then I ran the existing benchmarks before and after the changes, to see if using a function reference instead of directly invoking `maphash.Comparable` and `==` would have an impact and as expected, it does (see results below). My guess is that the primary reason this doesn't perform as well is that things aren't inlined.

Yes, it's a perf hit, but the question is whether it's significant enough to *not* introduce this feature, as there are few to no map implementations out there that allow using `[]byte` as the key, or any other non-`comparable` type.

```
% benchstat old=_sandbox/all new=_sandbox/new
goos: linux
goarch: amd64
pkg: github.com/puzpuzpuz/xsync/v4
cpu: AMD EPYC 7763 64-Core Processor
                                    │     old      │                  new                  │
                                    │    sec/op    │    sec/op      vs base                │
Map_NoWarmUp/reads=99%-8              8.001n ±  4%    8.957n ± 26%  +11.96% (p=0.000 n=10)
Map_NoWarmUp/reads=90%-8              10.60n ±  2%    14.08n ± 24%  +32.83% (p=0.000 n=10)
Map_NoWarmUp/reads=75%-8              14.06n ±  2%    18.00n ±  2%  +27.94% (p=0.000 n=10)
Map_WarmUp/reads=100%-8               8.521n ±  1%   10.140n ± 14%  +18.99% (p=0.000 n=10)
Map_WarmUp/reads=99%-8                7.684n ±  1%    8.188n ±  1%   +6.55% (p=0.000 n=10)
Map_WarmUp/reads=90%-8                9.638n ±  1%   10.140n ±  1%   +5.21% (p=0.000 n=10)
Map_WarmUp/reads=75%-8                12.38n ±  3%    12.86n ±  2%   +3.88% (p=0.002 n=10)
MapInt_NoWarmUp/reads=99%-8           6.660n ±  3%    7.263n ±  3%   +9.05% (p=0.000 n=10)
MapInt_NoWarmUp/reads=90%-8           9.255n ±  6%    9.692n ±  6%   +4.72% (p=0.001 n=10)
MapInt_NoWarmUp/reads=75%-8           12.07n ±  2%    12.49n ±  4%   +3.48% (p=0.005 n=10)
MapInt_WarmUp/reads=100%-8            5.335n ±  0%    6.005n ±  1%  +12.57% (p=0.000 n=10)
MapInt_WarmUp/reads=99%-8             6.371n ±  3%    6.803n ±  0%   +6.79% (p=0.000 n=10)
MapInt_WarmUp/reads=90%-8             8.087n ±  1%    8.566n ±  2%   +5.92% (p=0.000 n=10)
MapInt_WarmUp/reads=75%-8             10.55n ±  1%    10.96n ±  2%   +3.89% (p=0.000 n=10)
IntMapStandard_NoWarmUp/reads=99%-8   8.846n ± 11%    8.869n ±  2%        ~ (p=0.853 n=10)
IntMapStandard_NoWarmUp/reads=90%-8   12.03n ±  2%    12.06n ±  3%        ~ (p=0.670 n=10)
IntMapStandard_NoWarmUp/reads=75%-8   16.60n ±  2%    16.84n ±  1%        ~ (p=0.079 n=10)
IntMapStandard_WarmUp/reads=100%-8    7.905n ±  2%    7.904n ±  1%        ~ (p=0.971 n=10)
IntMapStandard_WarmUp/reads=99%-8     8.691n ±  1%    8.670n ±  1%        ~ (p=0.272 n=10)
IntMapStandard_WarmUp/reads=90%-8     11.78n ±  2%    11.99n ±  3%        ~ (p=0.055 n=10)
IntMapStandard_WarmUp/reads=75%-8     16.50n ±  2%    16.62n ±  2%        ~ (p=0.239 n=10)
MapRange-8                            3.127µ ±  4%    3.041µ ±  3%        ~ (p=0.159 n=10)
Compute/op=UpdateOp-8                 27.39n ±  0%    30.46n ±  0%  +11.25% (p=0.000 n=10)
Compute/op=NoOp-8                     14.88n ±  0%    17.20n ±  1%  +15.59% (p=0.000 n=10)
geomean                               13.06n          14.01n         +7.26%

                                    │      old      │                 new                  │
                                    │     ops/s     │    ops/s      vs base                │
Map_NoWarmUp/reads=99%-8               125.0M ±  3%   111.7M ± 21%  -10.65% (p=0.000 n=10)
Map_NoWarmUp/reads=90%-8               94.35M ±  2%   71.03M ± 24%  -24.72% (p=0.000 n=10)
Map_NoWarmUp/reads=75%-8               71.10M ±  2%   55.57M ±  2%  -21.84% (p=0.000 n=10)
Map_WarmUp/reads=100%-8               117.35M ±  1%   98.59M ± 16%  -15.98% (p=0.000 n=10)
Map_WarmUp/reads=99%-8                 130.1M ±  1%   122.1M ±  1%   -6.14% (p=0.000 n=10)
Map_WarmUp/reads=90%-8                103.76M ±  1%   98.64M ±  1%   -4.94% (p=0.000 n=10)
Map_WarmUp/reads=75%-8                 80.76M ±  3%   77.72M ±  2%   -3.77% (p=0.002 n=10)
MapInt_NoWarmUp/reads=99%-8            150.1M ±  2%   137.7M ±  3%   -8.27% (p=0.000 n=10)
MapInt_NoWarmUp/reads=90%-8            108.1M ±  7%   103.2M ±  6%   -4.50% (p=0.001 n=10)
MapInt_NoWarmUp/reads=75%-8            82.86M ±  2%   80.11M ±  4%   -3.32% (p=0.005 n=10)
MapInt_WarmUp/reads=100%-8             187.4M ±  0%   166.5M ±  1%  -11.17% (p=0.000 n=10)
MapInt_WarmUp/reads=99%-8              157.0M ±  3%   147.0M ±  0%   -6.36% (p=0.000 n=10)
MapInt_WarmUp/reads=90%-8              123.7M ±  1%   116.7M ±  2%   -5.60% (p=0.000 n=10)
MapInt_WarmUp/reads=75%-8              94.80M ±  1%   91.25M ±  2%   -3.74% (p=0.000 n=10)
IntMapStandard_NoWarmUp/reads=99%-8    113.1M ± 10%   112.8M ±  2%        ~ (p=0.853 n=10)
IntMapStandard_NoWarmUp/reads=90%-8    83.10M ±  2%   82.89M ±  3%        ~ (p=0.684 n=10)
IntMapStandard_NoWarmUp/reads=75%-8    60.25M ±  2%   59.40M ±  1%        ~ (p=0.075 n=10)
IntMapStandard_WarmUp/reads=100%-8     126.5M ±  2%   126.5M ±  1%        ~ (p=0.971 n=10)
IntMapStandard_WarmUp/reads=99%-8      115.1M ±  1%   115.3M ±  1%        ~ (p=0.280 n=10)
IntMapStandard_WarmUp/reads=90%-8      84.90M ±  2%   83.43M ±  3%        ~ (p=0.063 n=10)
IntMapStandard_WarmUp/reads=75%-8      60.59M ±  2%   60.17M ±  2%        ~ (p=0.247 n=10)
MapRange-8                             319.9k ±  4%   328.9k ±  3%        ~ (p=0.165 n=10)
geomean                                79.67M         74.66M         -6.29%

                                    │      old       │                 new                  │
                                    │      B/op      │    B/op      vs base                 │
Map_NoWarmUp/reads=99%-8              0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
Map_NoWarmUp/reads=90%-8              1.000 ±   0%     1.000 ±  0%       ~ (p=1.000 n=10) ¹
Map_NoWarmUp/reads=75%-8              3.000 ±  33%     3.000 ± 33%       ~ (p=0.628 n=10)
Map_WarmUp/reads=100%-8               0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=99%-8                0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=90%-8                1.000 ±   0%     1.000 ±  0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=75%-8                2.500 ±  20%     2.500 ± 20%       ~ (p=1.000 n=10)
MapInt_NoWarmUp/reads=99%-8           0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
MapInt_NoWarmUp/reads=90%-8           0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
MapInt_NoWarmUp/reads=75%-8           2.000 ±   0%     2.000 ± 50%       ~ (p=1.000 n=10)
MapInt_WarmUp/reads=100%-8            0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=99%-8             0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=90%-8             0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=75%-8             1.000 ± 100%     2.000 ± 50%       ~ (p=0.370 n=10)
IntMapStandard_NoWarmUp/reads=99%-8   0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_NoWarmUp/reads=90%-8   3.000 ±   0%     3.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_NoWarmUp/reads=75%-8   8.000 ±   0%     8.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=100%-8    0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=99%-8     0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=90%-8     3.000 ±   0%     3.000 ±  0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=75%-8     8.000 ±   0%     8.000 ±  0%       ~ (p=1.000 n=10) ¹
MapRange-8                            0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
Compute/op=UpdateOp-8                 1.000 ±   0%     1.000 ±  0%       ~ (p=1.000 n=10) ¹
Compute/op=NoOp-8                     0.000 ±   0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
geomean                                            ²                +2.93%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                    │     old      │                 new                 │
                                    │  allocs/op   │ allocs/op   vs base                 │
Map_NoWarmUp/reads=99%-8              0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_NoWarmUp/reads=90%-8              0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_NoWarmUp/reads=75%-8              0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=100%-8               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=99%-8                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=90%-8                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Map_WarmUp/reads=75%-8                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_NoWarmUp/reads=99%-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_NoWarmUp/reads=90%-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_NoWarmUp/reads=75%-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=100%-8            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=99%-8             0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=90%-8             0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapInt_WarmUp/reads=75%-8             0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_NoWarmUp/reads=99%-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_NoWarmUp/reads=90%-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_NoWarmUp/reads=75%-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=100%-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=99%-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=90%-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
IntMapStandard_WarmUp/reads=75%-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
MapRange-8                            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Compute/op=UpdateOp-8                 1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
Compute/op=NoOp-8                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                          ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```